test(ci): split test_validation — pure-logic classes into tests/unit (#1895)
Batch 3c. test_validation.py is mixed: 5 pure-logic classes
(TestValidationPromptBuilding / ...ResponseParsing / BusinessStatusMapping /
TestDatabaseOperations / TestScheduleValidationConfig) that stub the backend at
module import, plus TestValidationIntegration whose methods take api_client /
created_agent.

- Move the 5 pure-logic classes -> tests/unit/test_validation.py (they carry all 4
  module-level sys.modules stubs; baseline line relocated to the unit path).
  sys.path.insert base (Case A) — no re-anchor. Verified: --co 19, 16 passed / 3
  skipped under seeds 12345/67890/99999.
- TestValidationIntegration stays in tests/ root (auto-exempt: it requests the
  live api_client/created_agent fixtures). 0 sys.modules findings there.

Deviation (Attempt-3): test_ip_rate_limit_fix.py is NOT split — the whole file
(unit tests included) is red-and-hidden. _load_public_router imports routers.public,
which now does `from dependencies import ..., assert_owns` (#1310); the fixture's
MagicMock `dependencies` stub raises on an assert_*-named attribute (mock's
assertion-name guard), so all 14 tests error identically on origin/dev. It stays put
and is baselined by the new placement guard.
